requirements.txt
contourpy==1.3.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.57.0
fsspec==2024.6.1
imageio==2.37.0
importlib_resources==6.5.2
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.2.1
numpy==1.23.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
opencv-python==4.5.5.64
packaging==24.2
Pillow==9.5.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scipy==1.11.4
six==1.17.0
sympy==1.13.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.65.0
triton==3.1.0
typing_extensions==4.12.2
zipp==3.21.0
